Move files to config folder
