+++ extensions/ql-vscode/src/databases/config/db-config-validator.ts
