+++ extensions/ql-vscode/src/databases/db-manager.ts
@@ -1,7 +1,7 @@
 import { App } from '../common/app';
 import { AppEvent, AppEventEmitter } from '../common/events';
 import { ValueResult } from '../common/value-result';
-import { DbConfigStore } from './db-config-store';
+import { DbConfigStore } from './config/db-config-store';
 import { DbItem } from './db-item';
 import { createLocalTree, createRemoteTree } from './db-tree-creator';
