+++ extensions/ql-vscode/src/vscode-tests/minimal-workspace/databases/db-panel.test.ts
@@ -3,9 +3,9 @@ import { expect } from 'chai';
 import * as path from 'path';
 import * as fs from 'fs-extra';
 import * as pq from 'proxyquire';
-import { DbConfig } from '../../../databases/db-config';
+import { DbConfig } from '../../../databases/config/db-config';
 import { DbManager } from '../../../databases/db-manager';
-import { DbConfigStore } from '../../../databases/db-config-store';
+import { DbConfigStore } from '../../../databases/config/db-config-store';
 import { DbTreeDataProvider } from '../../../databases/ui/db-tree-data-provider';
 import { DbPanel } from '../../../databases/ui/db-panel';
 import { DbItemKind } from '../../../databases/db-item';
