+++ extensions/ql-vscode/test/pure-tests/databases/config/data/without-selected/workspace-databases.json
