+++ extensions/ql-vscode/test/pure-tests/databases/config/data/workspace-databases.json
