+++ extensions/ql-vscode/test/tsconfig.json
@@ -5,6 +5,6 @@
   "compilerOptions": {
     "noEmit": true,
     "resolveJsonModule": true,
-    "rootDirs": [".", "../src"]
+    "rootDir": ".."
   }
 }
